Federated_learning/requirements.txt

torch
torchvision
numpy
Pillow
fastapi
uvicorn[standard] 
requests         
tqdm   
python-multipart      